+++ sigle/src/modules/publicHome/components/PublicHome.tsx
@@ -6,6 +6,7 @@ import { StoryFile, SettingsFile } from '../../../types';
 import { Container } from '../../../components';
 import { PublicStoryItem } from './PublicStoryItem';
 import { PoweredBy } from '../../publicStory/PoweredBy';
+import { AppHeader } from '../../layout/components/AppHeader';
 
 const StyledContainer = styled(Container)`
   ${tw`pt-4 pb-16`};
@@ -52,6 +53,7 @@ export const PublicHome = ({ file, settings }: PublicHomeProps) => {
   return (
     <React.Fragment>
       <Container>
+        <AppHeader />
         <Header>
           {settings.siteLogo && (
             <HeaderLogo src={settings.siteLogo} alt={`${siteName} logo`} />